Change byte for long

+++ src/Main.java
@@ -6,9 +6,9 @@ public class Main {
         Scanner s = new Scanner(System.in);
         String number1 = s.next();
         String number2 = s.next();
-        Byte num1 = Byte.valueOf(number1);
-        Byte num2 = Byte.valueOf(number2);
-            if (num1 > num2){
+        Long num1 = Long.valueOf(number1);
+        Long num2 = Long.valueOf(number2);
+        if (num1 > num2){
                 System.out.println("Махатма");
             } else {
                 System.out.println("Джавахарлал");
